TP2-parte3/04_Otros/requirements.txt
pillow==11.0.0
tk==0.1.0
